Refactor StoreSDK to be accessible to other modules. Refactor to allow Products to be purchased on Nightly and Beta.
--

Fix OrderID coming back as an error message sometimes instead of empty.
Remove for loop running the backend processing code 10 times in a row. With Backend's multi-device processing, it should NOT be an issue!
With our current Multi-Device processing on device, it should automatically try again at some Apple set interval on ANY device.

--

Allow VPN to work on Nightly/Beta
