+++ ios/brave-ios/App/Configuration/Base.xcconfig
@@ -18,3 +18,6 @@ BASE_BUNDLE_ID = $(FRAMEWORK_BASE_BUNDLE_ID).ios
 DEVELOPMENT_TEAM = KL8N8XSYF4
 
 BRAVE_URL_SCHEME = brave
+
+VPN_PRODUCT_ID_PREFIX = bravevpn
+LEO_PRODUCT_ID_PREFIX = braveleo
