+++ ios/brave-ios/App/Configuration/Debug-AppStore.xcconfig
@@ -11,3 +11,6 @@ BRAVE_GROUP_ID = group.$(MOZ_BUNDLE_ID)
 // Manual Code-Signing
 DEVELOPMENT_TEAM = KL8N8XSYF4
 CODE_SIGN_STYLE = Manual
+
+VPN_PRODUCT_ID_PREFIX = bravevpn
+LEO_PRODUCT_ID_PREFIX = braveleo
