+++ ios/brave-ios/App/Configuration/Nightly.xcconfig
@@ -25,3 +25,6 @@ GCC_PREPROCESSOR_DEFINITIONS= DEBUG=1
 // Manual Code-Signing
 CODE_SIGN_IDENTITY = Apple Distribution: Brave Software, Inc.
 CODE_SIGN_STYLE = Manual
+
+VPN_PRODUCT_ID_PREFIX = nightly.bravevpn
+LEO_PRODUCT_ID_PREFIX = nightly.braveleo
