+++ ios/brave-ios/Sources/AIChat/Components/Debug/AIChatLeoSkusLogsView.swift
@@ -9,6 +9,7 @@ import OSLog
 import Preferences
 import SwiftUI
 import os.log
+import BraveStore
 
 public struct AIChatLeoSkusLogsView: View {
   @State
