+++ ios/brave-ios/Sources/Brave/BraveSkus/BraveSkusAccountLink.swift
@@ -4,6 +4,7 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 import AIChat
+import BraveStore
 import Foundation
 import Preferences
 import WebKit
